Create Phase1.java
modified as follows
i changed some imported package (from line 31 to 39)
2 functions are added
1. (from line 303 to 352)
addInfo: create and open the new database called Info. This method is called from writing method (line 409). This database contains 5 columns, the order is url, title, date, size_byte, size_char. If want to change the path of the db, please change on the line 320. After this method is finished, the database of Info will close.

2. from line 354 to 407
This method is not test / called by others yet. However, the content is similar to the method addInfo above.

Remark: the date present here is not the same as previous we print out. If want to change the format, please adjust on line 324 and line 377.

----------------------------------------------------
Other minor changes:
In the method writing(...), I modify the line 442 and 443 which was  deciding what format of size should print out and now is to call the addInfo method.

To be easier to observe what I was doing, I make some lines (originally printing out) into comment.